refactor tsdbtest package to separate schema creation logic to eliminate cyclic imports in partition manager test
