+++ test/benchmark/BenchmarkIngestWithNuclio_test.go
@@ -9,6 +9,7 @@ import (
 	"github.com/v3io/v3io-go-http"
 	"github.com/v3io/v3io-tsdb/pkg/tsdb"
 	"github.com/v3io/v3io-tsdb/pkg/tsdb/tsdbtest"
+	"github.com/v3io/v3io-tsdb/pkg/tsdb/tsdbtest/testutils"
 	"github.com/v3io/v3io-tsdb/pkg/utils"
 	"github.com/v3io/v3io-tsdb/test/benchmark/common"
 	"io/ioutil"
@@ -41,7 +42,7 @@ func BenchmarkIngestWithNuclio(b *testing.B) {
 
 	// Update TSDB instance path for this test
 	v3ioConfig.Path = tsdbPath
-	schema := tsdbtest.CreateSchema(b, "count,sum")
+	schema := testutils.CreateSchema(b, "count,sum")
 	if err := tsdb.CreateTSDB(v3ioConfig, &schema); err != nil {
 		b.Fatal("Failed to create TSDB", err)
 	}
